Adicionadas 2 instrucoes do 6502, e mais duas outras para testar...

+++ target-6502/cpu.h
@@ -254,13 +254,13 @@ typedef struct CPU6502State {
 
 struct CPUAlphaState {
     //      "General" Registers (they are really 8 bit, but as TCG doesn't
-    // seem to have 8 bit registers, we are going to use 8 bits...
-    uint32_t    ac;
-    uint32_t    x;
-    uint32_t    y;
+    // seem to have 8 bit registers, we are going to use more bits...
+    uint64_t    ac;
+    uint64_t    x;
+    uint64_t    y;
 
-    uint32_t    sr;     // These are the flags: NV-BDIZC
-    uint32_t    sp;
+    uint64_t    sr;     // These are the flags: NV-BDIZC
+    uint64_t    sp;
 
     uint64_t    pc;     // This is temporarily 64 bits so that we don't break the remaining Alpha code...
